Small refactor of Navigation component
